There is now a countdown!
